Extract `lib.user.user_sheet`
